+++ kamehouse-shell/bin/kamehouse/docker/docker-build-push-all-kamehouse.sh
@@ -11,6 +11,7 @@ initScriptEnv() {
 }
 
 mainProcess() {
+  checkDockerScripsEnabled
   setKameHouseRootProjectDir
   ${HOME}/programs/kamehouse-shell/bin/kamehouse/docker/docker-build-kamehouse.sh -b
   checkCommandStatus "$?" "Error rebuilding and pushing the kamehouse docker image"