+++ kamehouse-shell/bin/kamehouse/docker/docker-reinit-container.sh
@@ -17,6 +17,7 @@ initScriptEnv() {
 }
 
 mainProcess() {
+  checkDockerScripsEnabled
   printReinitSettings
 
   requestConfirmation