+++ kamehouse-shell/bin/kamehouse/docker/docker-run-kamehouse.sh
@@ -24,6 +24,7 @@ initScriptEnv() {
 }
 
 mainProcess() {
+  checkDockerScripsEnabled
   setEnvironment
   printEnv
   runDockerImage